refactor: ♻️ move app.data.encoding module to app.core.encoding and update import statements
