+++ app/docs/dependencies/custom_parameters.py
@@ -1,6 +1,6 @@
 from app.config.api_settings import get_settings
 from app.core.cache import cached_data
-from app.data.encoding import get_uri_encoded_value
+from app.core.encoding import get_uri_encoded_value
 from app.docs.dependencies import (
     BIDI_CLASS_VALUES_TABLE,
     BLOCK_NAME_NO_LEGEND_TABLE,
