+++ app/tests/test_character_endpoints/test_get_unicode_character_details/test_get_character_details.py
@@ -4,7 +4,7 @@ from functools import reduce
 import pytest
 
 from app.core.cache import cached_data
-from app.data.encoding import get_uri_encoded_value
+from app.core.encoding import get_uri_encoded_value
 from app.db.character_props import PROPERTY_GROUPS
 from app.db.procs.get_char_details import get_prop_groups
 from app.schemas.enums import CharPropertyGroup
